1.Fix the build error caused by the modification for compiler 6 2.Modify settings of Keil project to speed up compilation. 3.Revise copyright format for gen. chm file. 4.Remove useless files of IAR Project. 5.Fix download issue of IAR Project.

+++ SampleCode/StdDriver/ADC_PwmTrigger/main.c
@@ -6,8 +6,9 @@
  * @brief    NUC029 Series ADC Interface Controller Driver Sample Code
  *
  * @note
- * Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
+ * @copyright SPDX-License-Identifier: Apache-2.0
  *
+ * @copyright Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
  ******************************************************************************/
 #include <stdio.h>
 #include "NUC029xAN.h"
@@ -159,7 +160,7 @@ void ADC_PWMTrigTest_SingleOpMode()
     /* Clear the ADC interrupt flag */
     ADC_CLR_INT_FLAG(ADC, ADC_ADF_INT);
 
-    printf("Channel 2: 0x%X\n", ADC_GET_CONVERSION_DATA(ADC, 2));
+    printf("Channel 2: 0x%X\n",(unsigned int) ADC_GET_CONVERSION_DATA(ADC, 2));
 
     /* Disable ADC */
     ADC_POWER_DOWN(ADC);
